+++ CHANGELOG.md
@@ -1,6 +1,22 @@
+## 1.7.0 (November 15, 2016)
+BREAKING CHANGES:
+  - **Dropped support of Parallels Desktop 8 and 9**. These versions have 
+  reached their [End-of-Life and End-of-Support](http://kb.parallels.com/eu/122533).
+  - **Removed customization options, which were previously deprecated:**
+  [[GH-271](https://github.com/Parallels/vagrant-parallels/pull/271)]
+    - "use_linked_clone" - use `linked_clone` instead.
+    - "regen_box_uuid" - use `regen_src_uuid` instead.
+    - "optimize_power_consumption".  
+  
+FEATURES:
+  - **IPv6 Private Networks:** Private networking now supports IPv6. 
+  This only works with Parallels Desktop 12 and higher.
+  [[GH-273](https://github.com/Parallels/vagrant-parallels/pull/273)]
+
+
 ## 1.6.3 (July 11, 2016)
 DEPRECATIONS:
-  - The following provider options was renamed:
+  - The following provider options were renamed:
     - `regen_box_uuid` was renamed to `regen_src_uuid`
     - `use_linked_clone` was renamed to `linked clone`